Conflicts:
	arangod/Replication/Syncer.cpp
	arangod/VocBase/index-garbage-collector.c
	build_posix.h
	html/admin/api-docs/traversal.json
	js/apps/aardvark/api-docs/collection.json
	js/apps/aardvark/api-docs/cursor.json
	js/apps/aardvark/api-docs/document.json
	js/apps/aardvark/api-docs/edge.json
	js/apps/aardvark/api-docs/edges.json
	js/apps/aardvark/api-docs/graph.json
	js/apps/aardvark/api-docs/index.json
	js/apps/aardvark/api-docs/replication.json
	js/apps/aardvark/api-docs/simple.json
	js/apps/aardvark/api-docs/system.json

+++ lib/Admin/RestAdminLogHandler.cpp
@@ -53,7 +53,7 @@ static int LidCompareAsc (void const* l, void const* r) {
   TRI_log_buffer_t const* left = (TRI_log_buffer_t const*) l;
   TRI_log_buffer_t const* right = (TRI_log_buffer_t const*) r;
 
-  return ((int64_t) left->_lid) - ((int64_t) right->_lid);
+  return (int) (((int64_t) left->_lid) - ((int64_t) right->_lid));
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -64,7 +64,7 @@ static int LidCompareDesc (void const* l, void const* r) {
   TRI_log_buffer_t const* left = (TRI_log_buffer_t const*) l;
   TRI_log_buffer_t const* right = (TRI_log_buffer_t const*) r;
 
-  return ((int64_t) right->_lid) - ((int64_t) left->_lid);
+  return (int) (((int64_t) right->_lid) - ((int64_t) left->_lid));
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -288,8 +288,7 @@ HttpHandler::status_e RestAdminLogHandler::execute () {
   TRI_vector_t * logs = TRI_BufferLogging(ul, start, useUpto);
 
   if (logs == 0) {
-    generateError(HttpResponse::SERVER_ERROR,
-                  TRI_ERROR_OUT_OF_MEMORY);
+    generateError(HttpResponse::SERVER_ERROR, TRI_ERROR_OUT_OF_MEMORY);
     return HANDLER_DONE;
   }
